Refactored the algorithm to remove some duplication
The pattern that was obvious was that from 5 onwards, there was a new
starting point or symbol, ie V. So by checking if the amount is 5 or higher
I could start by having V already and then letting the algorithm do the
rest. What I still don't like is the 4 case but what can you do.